+++ Npm/contracts/queries/Operator.ts
@@ -1,7 +1,0 @@
-export const Equal = '='
-export const NotEqual = '<>'
-
-/**
- * Defines the type of operators supported in Queries Where conditions.
- */
-export type Operator = typeof Equal | typeof NotEqual