+++ Npm/index.ts
@@ -1,3 +1,3 @@
-export * from './contracts'
-export * from './components'
-export * from './database'
+export * from './src/contracts'
+export * from './src/components'
+export * from './src/database'