+++ RealTime/Queries/Condition.cs
@@ -19,7 +19,7 @@ namespace QualityGate.RealTime.Queries
         ///     The value that in order for the condition to be true, must be compared to the domain entity's field value
         ///     using the specified logical operator.
         /// </param>
-        public Condition(string field, OperatorBase @operator, object? value = null)
+        public Condition(string field, Operator @operator, object? value = null)
         {
             Field = field;
             Operator = @operator;
@@ -38,8 +38,10 @@ namespace QualityGate.RealTime.Queries
         {
             var condition = new Condition(conditionDto.Field, conditionDto.Operator, conditionDto.Value);
 
-            if (conditionDto.JoinUsing is not null)
-                condition.JoinUsing = conditionDto.JoinUsing;
+            if (conditionDto.JoinUsing is not null)  condition.JoinUsing = conditionDto.JoinUsing;
+            
+            condition.LeftParenthesis = conditionDto.LeftParenthesis;
+            condition.RightParenthesis = conditionDto.RightParenthesis;
 
             return condition;
         }
@@ -55,9 +57,19 @@ namespace QualityGate.RealTime.Queries
         public string Field { get; }
 
         /// <summary>
+        ///     Gets or sets an optional value saying whether or not this condition has a parenthesis to its left.
+        /// </summary>
+        public bool? LeftParenthesis { get; set; }
+
+        /// <summary>
         ///     Gets the logical operator to use in the comparison.
         /// </summary>
-        public OperatorBase Operator { get; }
+        public Operator Operator { get; }
+
+        /// <summary>
+        ///     Gets or sets an optional value saying whether or not this condition has a parenthesis to its right.
+        /// </summary>
+        public bool? RightParenthesis { get; set; }
 
         /// <summary>
         ///     Gets the value to compare the domain entity field with.
@@ -78,7 +90,7 @@ namespace QualityGate.RealTime.Queries
             var propertyInfo = entity.GetType().GetProperty(Field);
             var propertyValue = propertyInfo?.GetValue(entity);
 
-            return OperatorBase.Equal.Evaluate(Value, propertyValue);
+            return Operator.Equal.Evaluate(Value, propertyValue);
         }
 
         /// <summary>
@@ -95,7 +107,10 @@ namespace QualityGate.RealTime.Queries
                 null => "null",
                 _ => $"{Value}"
             };
-            builder.Append($"{Field} {Operator.Sign} {value}");
+            var leftParenthesis = LeftParenthesis ?? false ? "(" : string.Empty;
+            var rightParenthesis = RightParenthesis ?? false ? ")" : string.Empty;
+            var statement = $"{leftParenthesis}{Operator.ToRql(Field, value)}{rightParenthesis}";
+            builder.Append(statement);
 
             if (JoinUsing is not null) builder.Append($" {JoinUsing.Operator}");