+++ RealTime/Queries/JoinOperator.cs
@@ -23,7 +23,14 @@ namespace QualityGate.RealTime.Queries
         ///     <see cref="JoinOperator"/> equivalent.
         /// </summary>
         /// <param name="operator">String representation of the RQL logical operator to convert.</param>
-        /// <returns>The converter RQL logical operator in <see cref="JoinOperator"/> format.</returns>
+        /// <returns>The converted RQL logical operator in <see cref="JoinOperator"/> format.</returns>
         public static implicit operator JoinOperator(string @operator) => new(@operator);
+
+        /// <summary>
+        ///     Implicitly converts a <see cref="JoinOperator"/> to its string representation.
+        /// </summary>
+        /// <param name="operator">String representation of the RQL logical operator to convert.</param>
+        /// <returns>The string representation of the given <see cref="JoinOperator"/>.</returns>
+        public static implicit operator string(JoinOperator @operator) => @operator.Operator;
     }
 }